Extract "ModelEvaluation" component
Also extract "ProgressRing" (to avoid circular dependency)
