+++ extensions/ql-vscode/src/view/model-editor/ModelEditorProgressRing.tsx
@@ -1,0 +1,8 @@
+import { VSCodeProgressRing } from "@vscode/webview-ui-toolkit/react";
+import { styled } from "styled-components";
+
+export const ModelEditorProgressRing = styled(VSCodeProgressRing)`
+  width: 16px;
+  height: 16px;
+  margin-right: 5px;
+`;
